backend/requirements.txt
click==8.1.2
Flask==2.1.1
Flask-MySQL==1.5.1
PyMySQL==0.10.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysqlclient==2.1.0
python-dotenv==0.20.0
Werkzeug==2.1.1
zipp==3.8.0
cryptography
Flask-Cors==3.0.10